requirements.txt
# ── Core Framework ───────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0          # Async SQLite driver (default; swap for asyncpg for PostgreSQL)
# asyncpg==0.29.0          # Uncomment for PostgreSQL

# ── Validation & Settings ─────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Security / Auth ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1             # passlib is incompatible with bcrypt 5.x
python-multipart==0.0.9    # Required for OAuth2 form-based login

# ── Logging ───────────────────────────────────────────────────────────────────
loguru==0.7.2

# ── Analytics / Data ──────────────────────────────────────────────────────────
python-dateutil==2.9.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0              # Async test client for FastAPI

# ── Export ────────────────────────────────────────────────────────────────────
# (csv is stdlib; json is stdlib — no extra deps needed)

# ── Dev Utilities ─────────────────────────────────────────────────────────────
python-dotenv==1.0.1
